# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server API for NodeList program",
        "title": "NodeList API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/v1",
    "paths": {
...
        "/node/amount": {
            "get": {
                "description": "Query the amount of the registered nodes in nodelist server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get the amount of all registered nodes",
                "responses": {
                    "200": {
                        "description": "return the nodes amount successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/node/info": {
            "get": {
                "description": "Query the information of all nodes, support paging",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get all nodes information(paging)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "paging start index (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of items to return per page(default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return node info list successfully",
                        "schema": {
                            "$ref": "#/definitions/server.NodeInfos"
                        }
                    },
                    "400": {
                        "description": "request parameter error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/node/info/delegation/{address}": {
            "get": {
                "description": "Query the information of the nodes that the specific owner has delegated his licenses to",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get the nodes information delegated by a specific owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address(an ethereum address with prefix '0x')",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return delegated nodes information successfully",
                        "schema": {
                            "$ref": "#/definitions/server.NodeInfos"
                        }
                    },
                    "400": {
                        "description": "request parameter error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/node/info/owner/{address}": {
            "get": {
                "description": "Query the node information through the owner address who owned the node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get the node information of the owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address(an ethereum address with prefix '0x')",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return the node information successfully",
                        "schema": {
                            "$ref": "#/definitions/server.NodeInfo"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/node/info/recipient/{address}": {
            "get": {
                "description": "Query the nodes information list through the recipient address who receives the node reward",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get the nodes information by the recipient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recipient address(an ethereum address with prefix '0x')",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "paging start index (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of items to return per page(default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return the nodes information list successfully",
                        "schema": {
                            "$ref": "#/definitions/server.NodeInfos"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.NodeInfo": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "commissionRate": {
                    "type": "integer"
                },
                "commissionRateLastModifyAt": {
                    "type": "string"
                },
                "delegationAmount": {
                    "type": "integer"
                },
                "delegationReward": {
                    "type": "string"
                },
                "nodeAddress": {
                    "type": "string"
                },
                "nodeID": {
                    "type": "integer"
                },
                "onlineDays": {
                    "type": "integer"
                },
                "onlineDays_RecentMonth": {
                    "type": "integer"
                },
                "onlineDays_RecentWeek": {
                    "type": "integer"
                },
                "recipient": {
                    "type": "string"
                },
                "registerDate": {
                    "type": "string"
                },
                "selfTotalReward": {
                    "type": "string"
                },
                "selfWithdrawedReward": {
                    "type": "string"
                }
            }
        },
        "server.NodeInfos": {
            "type": "object",
            "properties": {
                "infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.NodeInfo"
                    }
                }
            }
        },
...
    }
}